Some of the preliminary versions
Those are preliminary versions, each shows a slightly different approach, but basically do the same stuff.
V01 is not vectorized, so just made of slow for loops. Easier to understand, but slooooow.
V02 is the vectorized version, so for loops are now matrix operations.
V03 the filter code is converted into an object to avoid criticisms ;)
V04 a new mode "white" is added, not a good idea but looks nice.
V05 sensitivity to light is auto-adjusted to noise to avoid snow when nothing moves.
V06 a different transformation that is symmetrical for lower and higher than avg pixels, not so interersting.
V07 a rethink of the filter where odd-even frames are treated differently, but basically do the same stuff than simpler version V05

Consider V05 the "standard" version as for now.